+++ src/annotation.rs
@@ -1,6 +1,7 @@
 use serde::Deserialize;
 use serde_xml_rs::from_str;
 use serde_xml_rs::Error;
+use std::path::{Path, PathBuf};
 
 pub fn parse_xml(src: &str) -> Result<Annotation, Error> {
     from_str(src)
@@ -42,6 +43,14 @@ impl Annotation {
             }
             None => None,
         }
+    }
+
+    pub fn get_image_path(&self, data_dir: &Path, image_dir: &Option<PathBuf>) -> String {
+        let image_dir: String = match &image_dir {
+            Some(dir) => dir.to_str().unwrap().to_string(),
+            None => format!("{}/{}", data_dir.to_str().unwrap(), self.folder),
+        };
+        format!("{}/{}", image_dir, self.filename)
     }
 }