+++ ChangeLog.md
@@ -1,5 +1,6 @@
 2022-11
 
+- some general adjustments;  new --summary option
 - initial version
   - annotation xml parsing
   - image cropping